refactor(tokenizer,parser): WavCmdParams.value f64→u32 + from_text 返回 warnings
WavCmdParams 的 value 从 f64 改为 u32，贴合 BMS 规格中 #WAVCMD
的非负整数语义（音高 0-127、音量百分比、时长）。value 改为 u32
后补 derive Eq。

Bms::from_text 返回类型从 Result<Self, _> 改为
Result<(Self, Vec<ControlFlowWarning>), _>，传播控制流阶段的
非致命警告（如未闭合 #RANDOM 块）。方法无外部调用方，自由修改。

+++ bms/parser/src/audio.rs
@@ -7,14 +7,14 @@ use std::collections::BTreeMap;
 use bms_tokenizer::{BmsBase, BmsHeaderResDefAudio, WavIndex};
 
 /// `#WAVCMD` 的解析参数——音高/音量/时长覆盖。
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub struct WavCmdParams {
     /// 命令 ID（`00` = 音高、`01` = 音量、`02` = 时长）。
     pub command_id: String,
     /// 目标 WAV 索引。
     pub wav_index: String,
-    /// 参数值。
-    pub value: f64,
+    /// 参数值（非负整数）。
+    pub value: u32,
 }
 
 /// `#EXWAV` 的扩展音频效果参数。
